Add options for Makefile and auto.sh

+++ auto.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ "$1" == "" ]]; then
+	echo -e "Usage: $0 n[ormal]|e[vil]\n\tn - normal kernel (with code sign)\n\te - evil kernl (without code sign)"
+	exit 1
+fi
+
 echo "Rebuld the image..."
 rm -rf ./hd60M.img*
 printf "1\n\n\n60\nhd60M.img\n" | bximage > /dev/null
@@ -8,4 +13,8 @@ if [[ ! -d ./build/ ]]; then
 	mkdir build
 fi
 
-make
+if [[ "$1" == "n" ]]; then
+	make
+elif [[ "$1" == "e" ]]; then
+	make evil
+fi
